requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-genai==1.19.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.0
requests==2.31.0
python-dotenv==1.0.0
asyncio-throttle==1.0.2
httpx==0.25.2 